third_party/WebKit/LayoutTests/http/tests/devtools/profiler/cpu-profiler-profiling-expected.txt

Tests that CPU profiling works. http://crbug.com/52634

Profiler was enabled.

Running: testProfiling
found pageFunction cpu-profiler-profiling.js:16
found (anonymous) cpu-profiler-profiling.js:1

Profiler was disabled.
